Rails::Auth::ErrorPage::DebugMiddleware
This middleware provides detailed debugging information about a
request intended to simplify development, or for use debugging
authorization problems in a staging environment.

It displays what request was made, all of the credentials currently
in the Rack environment, and all of the candidate ACL entries
which match the current request.