Make allowed_by a mandatory argument of Rails::Auth.authorized!
I released 2.0.0 with this as an optional argument, but I really think
it should be mandatory.

Nobody is using 2.0.0 yet, so for maximum semantic versioning correctness
I could yank it to ensure this backward incompatible change is consistent
for all 2.x releases, or just leave it as is and tell people 2.0.0 is
"broken".

Upon updating the spec, I did discover Rails::Auth.authorized! is calling
Rails::Auth.set_allowed_by incorrectly.